Fix warnings

+++ src/c/core/session/stream/output_reliable_stream.c
@@ -273,12 +273,13 @@ bool on_full_output_buffer(ucdrBuffer* ub, void* args)
 
 uint16_t get_available_seq_num(uxrOutputReliableStream* stream)
 {
-    uint16_t used_blocks = uxr_seq_num_sub(stream->last_written, stream->last_acknown);
+    int used_blocks = uxr_seq_num_sub(stream->last_written, stream->last_acknown);
+
     used_blocks = (used_blocks == 0) ? 
                     0 : 
                     (used_blocks > stream->base.history) ? 
                         stream->base.history :
                         used_blocks - 1;
-    return (stream->base.history - used_blocks);
+    return (uint16_t)(stream->base.history - used_blocks);
 }
